+++ packages/firebase-utils/src/index.ts
@@ -1,3 +1,5 @@
 export * from './auth'
 export * from './firestore'
 export * from './storage'
+
+export * from './scripts'