+++ packages/firebase-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @yohs/firebase-utils
 
+## 0.2.0
+
+### Minor Changes
+
+- Add firebase-scripts helper
+
+## 0.1.3
+
+### Patch Changes
+
+- Fix collection iterate last doc index
+
 ## 0.1.2
 
 ### Patch Changes